Add a first draft of the dout command
This change introduces the dout command. The dcp.schema module has been
updated to include logic for actually visiting rows in the source
database (in the correct order) and some testing was added for verifying
its correctness.

The README has also been updated to begin explaining the logic used and
enumerating potential pitfalls.