Add memoization to enumerated data
As requests for data are passed to the schema object, the (table,
filters) pairs are now recorded to track what data has already been
requested. The testing for the extraction was also extended to include
data that would be returned multiple times if the memoization were not
in place to verify it is working correctly.